MapEditor(2)
- Editor Dialog 설계 완료
- BG 이미지 설정 설계중 (콤보박스에서 선택할 수 있도록만 만들어둠, 선택한 문자열로 텍스쳐 선택해서 버튼 누르면 실제 배경 적용되도록 구현해야 함)
- TIle편집 Editor 설계중 (원하는 팔레트 선택하면 텍스쳐를 띄우고, 거기서 원하는 부분 선택하면 그 타일을 메인 윈도우에서 마우스가 가리키는 타일에 반투명하게 렌더링 하는 데까지 설계함)

+++ CDeadEffect.h
@@ -11,7 +11,7 @@ private:
 
     float   m_Distance;
     float   m_MaxDistance;
-    float   m_Count;
+    int     m_Count;
 
     BANG_COLOR m_BaseColor;
     BANG_COLOR m_Color;
